Crate seed_riscv

Modules

block ğŸ”’

fs ğŸ”’

image ğŸ”’

logger ğŸ”’

memory ğŸ”’

pci ğŸ”’

Statics

ALLOCATOR ğŸ”’

MEMORY_MANAGER ğŸ”’

_bss_end ğŸ”’ âš

_bss_start ğŸ”’ âš

_seed_end ğŸ”’ âš

_seed_start ğŸ”’ âš

_stack_bottom ğŸ”’ âš

_stack_top ğŸ”’ âš

Functions

alloc_and_map_kernel_heap ğŸ”’

Allocate memory for the kernel heap, and dynamically map it into the address space after the kernel. We tell the kernel where to find it via the boot info.

create_boot_info ğŸ”’

Allocate memory for the boot info, and dynamically map it into the address space after the kernel.

seed_main

trap_handler