Struct seed_riscv::memory::MemoryManager

pub struct MemoryManager(Spinlock<MemoryManagerInner>);

The physical memory manager - this consumes a MemoryRegions map, and uses it to initialise an instrusive free list of all usable physical memory. This can then be used to allocate physical memory as needed, at frame granularity.

Tuple Fields
0: Spinlock<MemoryManagerInner>
Implementations

impl MemoryManager

pub const fn new() -> MemoryManager

pub fn init(&self, regions: MemoryRegions)

pub fn walk_usable_memory(&self)

pub fn populate_memory_map(&self, memory_map: &mut MemoryMap)